requirements.txt
fastapi>=0.115.0
uvicorn>=0.24.0
websockets>=12.0
pydantic>=2.4.2
python-dotenv>=1.0.0
crewai>=0.1.0
pandas>=2.0.0
pyyaml>=6.0
